Source/src/Tests/Promeetec.EDMS.Domain.Tests/Document/Rapportage/CommandHandlers/CreateRapportageHandlerTests.cs
using AutoFixture;
using FluentValidation;
using FluentValidation.Results;
using Microsoft.EntityFrameworkCore;
using Moq;
using NUnit.Framework;
using Promeetec.EDMS.Portaal.Data.Context;
using Promeetec.EDMS.Portaal.Data.Repositories;
using Promeetec.EDMS.Portaal.Domain.Models.Document.Rapportage;
using Promeetec.EDMS.Portaal.Domain.Models.Document.Rapportage.Commands;
using Promeetec.EDMS.Portaal.Domain.Models.Document.Rapportage.Handlers;
using Promeetec.EDMS.Portaal.Domain.Models.Event;
using Promeetec.EDMS.Portaal.Tests.Helpers;
using Promeetec.EDMS.Tests.Helpers;

namespace Promeetec.EDMS.Portaal.Domain.Tests.Document.Rapportage.CommandHandlers;



[TestFixture]
public class CreateOverigBestandHandlerTests : TestFixtureBase
{
    private EDMSDbContext _context;
    private IRapportageRepository _repository;
    private IEventRepository _eventRepository;

    [SetUp]
    public void Setup()
    {
        _context = new EDMSDbContext(Shared.CreateContextOptions());

        _repository = new RapportageRepository(_context);
        _eventRepository = new EventRepository(_context);
    }


    [Test]
    public async Task Should_create_new_overig_bestand_and_add_event()
    {
        var command = Fixture.Create<CreateRapportage>();

        var validator = new Mock<IValidator<CreateRapportage>>();
        validator.Setup(x => x.ValidateAsync(command, new CancellationToken())).ReturnsAsync(new ValidationResult());

        var sut = new CreateRapportageHandler(_repository, _eventRepository, validator.Object);

        await sut.Handle(command);

        var dbEntity = await _context.Rapportages.FirstOrDefaultAsync(x => x.Id == command.Id);
        var @event = await _context.Events.FirstOrDefaultAsync(x => x.TargetId == command.Id);

        validator.Verify(x => x.ValidateAsync(command, new CancellationToken()));
        Assert.NotNull(dbEntity);
        Assert.NotNull(@event);
    }
}